Move compile-ui-assets to `ui` command group in breeze.
There is a new `ui` command group and this command belongs there.
