--- dev/breeze/doc/images/output_compile-ui-assets.txt
@@ -1,1 +1,0 @@
-92830a5b5f1afbbd42017e2e4b7b2dbb
