+++ dev/breeze/doc/images/output_setup_regenerate-command-images.txt
@@ -1,1 +1,1 @@
-edd856a0a570e6ff1cb050d0177ea242
+d09390dba69a08bc690374089442def0
