+++ dev/breeze/doc/images/output_ui.txt
@@ -1,1 +1,1 @@
-1c824059bb3d2ae726b7ece89d6b13b7
+6ba040207b6f4706d1600f0fbb37e3cd
