+++ dev/breeze/doc/images/output_ui_compile-assets.txt
@@ -1,0 +1,1 @@
+60099c9883af325bb92a6313023d1c55
