+++ dev/breeze/src/airflow_breeze/commands/setup_commands.py
@@ -559,7 +559,6 @@ DEVELOPER_COMMANDS = [
     "exec",
     "shell",
     "run",
-    "compile-ui-assets",
     "cleanup",
     "generate-migration-file",
     "doctor",
